Updated Timer code to avoid a rare messaging problem. Fixed a couple of minor build errors. Rearranged the atomic functions and added a new compare-and-swap operation. Added a thread-priority tweak to WASAPI. Removed MS-specific classes from the web browser component.

+++ src/audio/plugins/formats/juce_AudioUnitPluginFormat.mm
@@ -351,8 +351,8 @@ private:
         {
             for (int i = 0; i < supportedChannelsSize / sizeof (AUChannelInfo); ++i)
             {
-                numIns = jmax (numIns, supportedChannels[i].inChannels);
-                numOuts = jmax (numOuts, supportedChannels[i].outChannels);
+                numIns = jmax (numIns, (int) supportedChannels[i].inChannels);
+                numOuts = jmax (numOuts, (int) supportedChannels[i].outChannels);
             }
         }
         else
